Move fake-atomic armv6m traits into lib crate.
