+++ lib/armv6m-atomic-hack/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "armv6m-atomic-hack"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[build-dependencies]
+build-util = {path = "../../build/util"}
