+++ task/hiffy/src/main.rs
@@ -20,30 +20,7 @@ use hif::*;
 use userlib::*;
 
 #[cfg(armv6m)]
-trait AtomicU32Ext {
-    fn swap(&self, val: u32, order: Ordering) -> u32;
-    fn fetch_add(&self, val: u32, order: Ordering) -> u32;
-    fn fetch_sub(&self, val: u32, order: Ordering) -> u32;
-}
-
-#[cfg(armv6m)]
-impl AtomicU32Ext for AtomicU32 {
-    fn swap(&self, val: u32, order: Ordering) -> u32 {
-        let rv = self.load(order);
-        self.store(val, order);
-        rv
-    }
-    fn fetch_add(&self, val: u32, order: Ordering) -> u32 {
-        let rv = self.load(order);
-        self.store(rv + val, order);
-        rv
-    }
-    fn fetch_sub(&self, val: u32, order: Ordering) -> u32 {
-        let rv = self.load(order);
-        self.store(rv - val, order);
-        rv
-    }
-}
+use armv6m_atomic_hack::AtomicU32Ext;
 
 mod common;
