+++ task/jefe/Cargo.toml
@@ -12,6 +12,7 @@ num-traits = { version = "0.2.12", default-features = false }
 zerocopy = "0.6.1"
 cortex-m = { version = "0.7", features = ["inline-asm"] }
 cortex-m-semihosting = { version = "0.3.7", features = ["inline-asm"], optional = true }
+armv6m-atomic-hack = {path = "../../lib/armv6m-atomic-hack"}
 
 [build-dependencies]
 build-util = {path = "../../build/util"}
